Responding to changes in PR.
